core/hasher.py
"""
core/hasher.py
--------------
Keyframe Extraction & Visual Embedding Engine.

Extracts keyframes from a video using scene-change detection (cv2.absdiff) 
to reduce redundant frames by 60-80%.
Generates semantic embeddings for each keyframe using CLIP (sentence-transformers).
"""

import cv2
import numpy as np
from PIL import Image
from pathlib import Path
from typing import Generator, Dict

from sentence_transformers import SentenceTransformer

# ─── Constants ────────────────────────────────────────────────────────────────

FRAME_INTERVAL_SEC: float = 2.0   # Baseline extraction fallback
SCENE_CHANGE_THRESHOLD: float = 30.0 # Absolute difference threshold for scene change
EMBEDDING_MODEL_NAME: str = "clip-ViT-B-32"

# ─── Global State ─────────────────────────────────────────────────────────────

_clip_model = None

def get_clip_model():
    global _clip_model
    if _clip_model is None:
        _clip_model = SentenceTransformer(EMBEDDING_MODEL_NAME)
    return _clip_model


# ─── Public API ───────────────────────────────────────────────────────────────

def extract_frames(
    video_path: str | Path,
    interval_sec: float = FRAME_INTERVAL_SEC,
    use_scene_detection: bool = True
) -> Generator[tuple[float, np.ndarray], None, None]:
    """
    Yield (timestamp_sec, frame_bgr) tuples from a video file.
    Uses OpenCV absolute difference to detect scene changes and return 
    only highly representative representative frames, skipping static / redundant ones.
    """
    cap = cv2.VideoCapture(str(video_path))
    if not cap.isOpened():
        if isinstance(video_path, str) and video_path.startswith("http"):
            cap = cv2.VideoCapture(str(video_path), cv2.CAP_FFMPEG)
            
    if not cap.isOpened():
        raise FileNotFoundError(f"Cannot open video source: {video_path}")

    fps = cap.get(cv2.CAP_PROP_FPS) or 30.0
    total_frames = cap.get(cv2.CAP_PROP_FRAME_COUNT)
    frame_step = max(1, int(fps * interval_sec))
    
    current_idx = 0
    prev_gray = None
    
    # Always yield the first frame
    ret, frame = cap.read()
    if ret and frame is not None:
        yield 0.0, frame
        prev_gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
        
    next_target = frame_step

    while True:
        ret = cap.grab()
        if not ret:
            break
            
        current_idx += 1
        
        if current_idx >= next_target:
            ret, frame = cap.retrieve()
            if not ret or frame is None:
                break

            timestamp = current_idx / fps
            
            if use_scene_detection and prev_gray is not None:
                gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
                # Calculate pixel absolute difference
                diff = cv2.absdiff(gray, prev_gray)
                mean_diff = np.mean(diff)
                
                # If the frame has changed sufficiently, yield it
                if mean_diff > SCENE_CHANGE_THRESHOLD:
                    yield timestamp, frame
                    prev_gray = gray
            else:
                yield timestamp, frame

            next_target += frame_step

    cap.release()


def preprocess_frame_for_hash(frame_bgr: np.ndarray) -> np.ndarray:
    """Strip black borders to make crop-invariance better for CLIP."""
    gray = cv2.cvtColor(frame_bgr, cv2.COLOR_BGR2GRAY)
    DARK = 20
    non_black_rows = np.any(gray > DARK, axis=1)
    non_black_cols = np.any(gray > DARK, axis=0)

    if not non_black_rows.any() or not non_black_cols.any():
        return frame_bgr

    r0, r1 = int(np.argmax(non_black_rows)), int(len(non_black_rows) - 1 - np.argmax(non_black_rows[::-1]))
    c0, c1 = int(np.argmax(non_black_cols)), int(len(non_black_cols) - 1 - np.argmax(non_black_cols[::-1]))

    h, w = frame_bgr.shape[:2]
    crop_h = r1 - r0 + 1
    crop_w = c1 - c0 + 1

    if crop_h > h * 0.50 and crop_w > w * 0.50:
        return frame_bgr[r0:r0 + crop_h, c0:c0 + crop_w]
    return frame_bgr


def generate_embeddings_batch(frames_bgr: list[np.ndarray]) -> np.ndarray:
    """Generate CLIP embeddings for a batch of frames utilizing native acceleration."""
    pil_imgs = []
    for frame in frames_bgr:
        frame_clean = preprocess_frame_for_hash(frame)
        rgb = cv2.cvtColor(frame_clean, cv2.COLOR_BGR2RGB)
        pil_imgs.append(Image.fromarray(rgb))
        
    model = get_clip_model()
    # Batch encode takes advantage of vectorized CPU instructions and GPU acceleration
    return model.encode(pil_imgs, batch_size=16)


def hash_video(
    video_path: str | Path,
    interval_sec: float = FRAME_INTERVAL_SEC
) -> Dict[str, np.ndarray]:
    """
    Generate CLIP embeddings for keyframes in a video.
    Returns: dict mapping "timestamp_sec" -> 512d numpy array.
    """
    results: Dict[str, np.ndarray] = {}
    
    # Collect frames for batched execution
    timestamps = []
    frames = []
    for timestamp, frame in extract_frames(video_path, interval_sec, use_scene_detection=True):
        timestamps.append(f"{timestamp:.2f}")
        frames.append(frame)
        
    if not frames:
        return results

    try:
        # Encode all at once using massive internal batch processing
        embeddings = generate_embeddings_batch(frames)
        for ts_key, emb in zip(timestamps, embeddings):
            results[ts_key] = emb
    except Exception as exc:
        print(f"Error computing CLIP embeddings batch: {exc}")

    return results

def hamming_distance(hash_a: str, hash_b: str) -> int:
    """Deprecated: Provided for backward compatibility. Use cosine similarity instead."""
    return 999
