requirements.txt
opencv-python>=4.8.0
imagehash>=4.3.1
Pillow>=10.0.0
ultralytics>=8.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
yt-dlp>=2023.12.30
beautifulsoup4>=4.12.0
httpx>=0.25.0
mysql-connector-python>=9.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
matplotlib>=3.7.0
reportlab>=4.0.0
pymongo>=4.5.0
duckduckgo_search>=4.0.0
python-dotenv
google-generativeai
